MC

Monte Carlo (MC)

This is plain simple Monte Carlo (MC). It can for example be used to simulate the Ising model (see 2D Ising model).

You can initialize a Monte Carlo simulation of a given model simply through

mc = MC(model)

Allowed keywords are:

beta: inverse temperature

sweeps: number of measurement sweeps

thermalization: number of thermalization (warmup) sweeps

global_moves: wether global moves should be proposed

global_rate: frequency for proposing global moves

seed: initialize MC with custom seed

Afterwards, you can run the simulation by

run!(mc)

Note that you can just do another run!(mc, sweeps=1000) to continue the simulation.

Exports

MonteCarlo.run! — Method.

run!(mc::MC[; verbose::Bool=true, sweeps::Int, thermalization::Int])

Runs the given Monte Carlo simulation mc. Progress will be printed to STDOUT if verbose=true (default).

source

MonteCarlo.MC — Type.

Monte Carlo simulation

source

MonteCarlo.MC — Method.

MC(m::M; kwargs::Dict{String, Any})

Create a Monte Carlo simulation for model m with (keyword) parameters as specified in the dictionary kwargs.

source

MonteCarlo.MC — Method.

MC(m::M; kwargs...) where M<:Model

Create a Monte Carlo simulation for model m with keyword parameters kwargs.

source

Model interface

Any model that wants to be simulatable by means of MC must implement the following interface.

Mandatory fields

l::Lattice: any Lattice

Mandatory methods

conftype: type of a configuration

energy: energy of configuration

rand: random configuration

propose_local: propose local move

accept_local!: accept a local move

Precise signatures can be found here: Interface: Monte Carlo (MC).

Optional methods

global_move: propose and accept or reject a local move

prepare_observables: initialize observables

measure_observables!: measure observables

finish_observables!: finish measurements

Precise signatures can be found here: Interface: Monte Carlo (MC).

Potential extensions

Pull requests are very much welcome!

Heat bath (instead of Metropolis) option