DQMC

Interface: Determinant Quantum Monte Carlo (DQMC)

Any model that wants to be simulated by means of DQMC must implement the following interface. Below you find all semantic definitions and precise signatures of mandatory fields and mandatory and optional methods that any model should implement to work with the determinant Monte Carlo flavor Determinant Quantum Monte Carlo (DQMC).

Example models: Attractive Hubbard model

Mandatory fields

l::Lattice: any Lattice

flv::Int: number of distinct fermion flavors (e.g. spins, bands, etc.). The Green's function will have shape (flv*N, flv*N), where N is the number of sites. Note that different fermion flavors are often related by symmetry and it's advisable to use this symmetry to work with smaller Green's function matrices. Have a look at the Attractive Hubbard Model as an example where flv=1 although we have spinful fermions.

Index of all methods

Base.Random.rand

MonteCarlo.accept_local

MonteCarlo.conftype

MonteCarlo.energy_boson

MonteCarlo.finish_observables!

MonteCarlo.greenseltype

MonteCarlo.hopping_matrix

MonteCarlo.interaction_matrix_exp!

MonteCarlo.measure_observables!

MonteCarlo.prepare_observables

MonteCarlo.propose_local

Mandatory methods

MonteCarlo.conftype — Method.

conftype(::Type{DQMC}, m::Model)

Returns the type of a configuration.

source

Base.Random.rand — Method.

rand(mc::DQMC, m::Model)

Draw random configuration.

source

MonteCarlo.hopping_matrix — Method.

hopping_matrix(mc::DQMC, m::Model)

Calculates the hopping matrix $ T_{i\sigma, j\sigma '} $ where $ i, j $ are site indices and $ \sigma , \sigma ' $ are flavor indices (e.g. spin indices). The hopping matrix should also contain potential chemical potential terms on the diagonal.

A matrix element is the hopping amplitude for a hopping process: $ j,\sigma ' \rightarrow i,\sigma $.

Regarding the order of indices, if T[i, σ, j, σ'] is your desired 4D hopping array, then reshape(T, (n_sites * n_flavors, :)) is the hopping matrix.

source

MonteCarlo.interaction_matrix_exp! — Function.

interaction_matrix_exp!(mc::DQMC, m::Model, result::Matrix, conf, slice::Int, power::Float64=1.) -> nothing

Calculate the interaction matrix exponential expV = exp(- power * delta_tau * V(slice)) and store it in result::Matrix. Potential chemical potential terms should be part of the hopping_matrix and not the interaction.

This is a performance critical method and one might consider efficient in-place (in result) construction.

source

MonteCarlo.propose_local — Method.

propose_local(mc::DQMC, m::Model, i::Int, conf, E_boson::Float64) -> detratio, delta_E_boson, delta

Propose a local move for lattice site i of current configuration conf with boson energy E_boson. Returns the Green's function determinant ratio, the boson energy difference delta_E_boson = delta_E_boson_new - delta_E_boson, and additional local move information delta (will be forwarded to accept_local!).

See also accept_local!.

source

MonteCarlo.accept_local — Method.

accept_local(mc::DQMC, m::Model, i::Int, slice::Int, conf, delta, detratio, delta_E_boson)

Accept a local move for site i at imaginary time slice slice of current configuration conf with energy E. Arguments delta, detratio and delta_E_boson correspond to output of propose_local() for that local move.

See also propose_local.

source

Optional methods

MonteCarlo.greenseltype — Method.

greenseltype(::Type{DQMC}, m::Model)

Returns the type of the elements of the Green's function matrix. Defaults to Complex128.

source

MonteCarlo.energy_boson — Method.

energy(mc::DQMC, m::Model, conf)

Calculate bosonic part (non-Green's function determinant part) of energy for configuration conf for Model m.

source

MonteCarlo.prepare_observables — Method.

prepare_observables(mc::DQMC, m::Model) -> Dict{String, Observable}

Initializes observables and returns a Dict{String, Observable}. In the latter, keys are abbreviations for the observables names and values are the observables themselves.

See also measure_observables! and finish_observables!.

source

MonteCarlo.measure_observables! — Method.

measure_observables!(mc::DQMC, m::Model, obs::Dict{String,Observable}, conf, E::Float64)

Measures observables and updates corresponding MonteCarloObservable.Observable objects in obs.

See also prepare_observables and finish_observables!.

source

MonteCarlo.finish_observables! — Method.

measure_observables!(mc::DQMC, m::Model, obs::Dict{String,Observable}, conf, E::Float64)

Measure observables and update corresponding MonteCarloObservable.Observable objects in obs.

See also prepare_observables and measure_observables!.

source