MC

Interface: Monte Carlo (MC)

Any model that wants to be simulated by means of MC must implement the following interface. Below you find all semantic definitions and precise signatures of mandatory fields and mandatory and optional methods that any model should implement to work with the Monte Carlo flavor Monte Carlo (MC).

Example models: Ising Model

Mandatory fields

l::Lattice: any Lattice

Index of all methods

Base.Random.rand

MonteCarlo.accept_local!

MonteCarlo.conftype

MonteCarlo.energy

MonteCarlo.finish_observables!

MonteCarlo.global_move

MonteCarlo.measure_observables!

MonteCarlo.prepare_observables

MonteCarlo.propose_local

Mandatory methods

MonteCarlo.conftype — Method.

conftype(::Type{MC}, m::Model)

Returns the type of a configuration.

source

MonteCarlo.energy — Method.

energy(mc::MC, m::Model, conf)

Calculate energy of configuration conf for Model m.

source

Base.Random.rand — Method.

rand(mc::MC, m::Model)

Draw random configuration.

source

MonteCarlo.propose_local — Method.

propose_local(mc::MC, m::Model, i::Int, conf, E::Float64) -> delta_E, delta_i

Propose a local move for lattice site i of current configuration conf with energy E. Returns local move information delta_i (e.g. new[i] - conf[i], will be forwarded to accept_local!) and energy difference delta_E = E_new - E_old.

See also accept_local!.

source

MonteCarlo.accept_local! — Method.

accept_local(mc::MC, m::Model, i::Int, conf, E::Float64, delta_i, delta_E::Float64)

Accept a local move for site i of current configuration conf with energy E. Arguments delta_i and delta_E correspond to output of propose_local() for that local move.

See also propose_local.

source

Optional methods

MonteCarlo.global_move — Method.

global_move(mc::MC, m::Model, conf, E::Float64) -> accepted::Bool

Propose a global move for configuration conf with energy E. Returns wether the global move has been accepted or not.

source

MonteCarlo.prepare_observables — Method.

prepare_observables(mc::MC, m::Model) -> Dict{String, Observable}

Initializes observables and returns a Dict{String, Observable}. In the latter, keys are abbreviations for the observables names and values are the observables themselves.

See also measure_observables! and finish_observables!.

source

MonteCarlo.measure_observables! — Method.

measure_observables!(mc::MC, m::Model, obs::Dict{String,Observable}, conf, E::Float64)

Measures observables and updates corresponding MonteCarloObservable.Observable objects in obs.

See also prepare_observables and finish_observables!.

source

MonteCarlo.finish_observables! — Method.

measure_observables!(mc::MC, m::Model, obs::Dict{String,Observable}, conf, E::Float64)

Measure observables and update corresponding MonteCarloObservable.Observable objects in obs.

See also prepare_observables and measure_observables!.

source