Methods: General

Below you find all general exports.

Index

MonteCarlo.AbstractLattice

MonteCarlo.Model

MonteCarlo.MonteCarloFlavor

Base.length

Base.size

MonteCarlo.lattice

MonteCarlo.neighbors

MonteCarlo.neighbors

MonteCarlo.parameters

MonteCarlo.reset!

MonteCarlo.run!

Documentation

Base.length — Method

length(l::AbstractLattice)

Number of lattice sites.

Base.size — Method

size(l::AbstractLattice)

Returns the linear system size of a lattice.

MonteCarlo.lattice — Method

lattice(model)

Returns the lattice of a given model.

MonteCarlo.neighbors — Method

neighbors(l::AbstractLattice, site_index::Integer)

Returns a list of site indices neighboring the specified site_index.

MonteCarlo.neighbors — Method

neighbors(l::AbstractLattice[, directed=Val(false)])

Returns an iterator over bonds, given as tuples (source index, target index). If directed = Val(true) bonds are assumed to be directed, i.e. both (1, 2) and (2, 1) are included. If directed = Val(false) bonds are assumed to be undirected, i.e. (1, 2) and (2, 1) are assumed to be equivalent and only one of them will be included.

MonteCarlo.parameters — Method

parameters(model)

Collects relevant parametrs of a model into a named tuple.

MonteCarlo.reset! — Method

reset!(mc::MonteCarloFlavor)

Resets the Monte Carlo simulation mc. Previously set parameters will be retained.

MonteCarlo.run! — Method

run!(mc)

Run the Monte Carlo Simulation.

MonteCarlo.AbstractLattice — Type

Abstract definition of a lattice.

MonteCarlo.Model — Type

Abstract model.

MonteCarlo.MonteCarloFlavor — Type

Abstract definition of a Monte Carlo flavor.