General

Methods: General

Below you find all general exports.

Index

MonteCarlo.IsingModel

MonteCarlo.IsingModel

MonteCarlo.IsingModel

MonteCarlo.MC

MonteCarlo.MC

MonteCarlo.MC

MonteCarlo.init!

MonteCarlo.run!

Documentation

MonteCarlo.init! — Method.

init!(mc::MC[; seed::Real=-1])

Initialize the Monte Carlo simulation mc. If seed !=- 1 the random generator will be initialized with srand(seed).

source

MonteCarlo.run! — Method.

run!(mc::MC[; verbose::Bool=true, sweeps::Int, thermalization::Int])

Runs the given Monte Carlo simulation mc. Progress will be printed to STDOUT if verborse=true (default).

source

MonteCarlo.MC — Type.

Monte Carlo simulation

source

MonteCarlo.MC — Method.

MC(m::M; kwargs::Dict{String, Any})

Create a Monte Carlo simulation for model m with (keyword) parameters as specified in the dictionary kwargs.

source

MonteCarlo.MC — Method.

MC(m::M; kwargs...) where M<:Model

Create a Monte Carlo simulation for model m with keyword parameters kwargs.

source

MonteCarlo.IsingModel — Type.

Famous Ising model on a cubic lattice.

source

MonteCarlo.IsingModel — Method.

IsingModel(kwargs::Dict{String, Any})

Create Ising model with (keyword) parameters as specified in kwargs dict.

source

MonteCarlo.IsingModel — Method.

IsingModel(; dims::Int=2, L::Int=8)

Create Ising model on dims-dimensional cubic lattice with linear system size L.

source