Models

Methods: Models

Below you find all methods that any particular implementation of the abstract type MonteCarlo.Model must implement.

Index

MonteCarlo.CubicLattice

MonteCarlo.Lattice

MonteCarlo.Model

MonteCarlo.MonteCarloFlavor

Base.Random.rand

MonteCarlo.accept_local!

MonteCarlo.conftype

MonteCarlo.energy

MonteCarlo.observables

MonteCarlo.propose_local

MonteCarlo.reset!

Documentation

MonteCarlo.observables — Method.

observables(m::Model)

Get all observables defined for a given model.

Returns a Dict{String, String} where values are the observables names and keys are short versions of those names. The keys can be used to collect correponding observable objects from a Monte Carlo simulation, e.g. like mc.obs[key].

source

Base.Random.rand — Method.

rand(m::Model)

Draw random configuration.

source

MonteCarlo.accept_local! — Method.

accept_local(m::Model, i::Int, conf, E::Float64)

Accept a local move for site i of current configuration conf with energy E. Arguments Δi and ΔE correspond to output of propose_local() for that local move.

source

MonteCarlo.conftype — Method.

conftype(m::Model)

Returns the type of a configuration.

source

MonteCarlo.energy — Method.

energy(m::Model, conf)

Calculate energy of configuration conf for Model m.

source

MonteCarlo.propose_local — Method.

propose_local(m::Model, i::Int, conf, E::Float64) -> ΔE, Δi

Propose a local move for element i of current configuration conf with energy E. Returns the local move Δi = new[i] - conf[i] and energy difference ΔE = E_new - E_old.

source

MonteCarlo.reset! — Method.

reset!(mc::MonteCarloFlavor)

Resets the Monte Carlo simulation mc. Previously set parameters will be retained.

source

MonteCarlo.CubicLattice — Type.

Abstract cubic lattice.

1D -> Chain

2D -> SquareLattice

ND -> NCubeLattice

source

MonteCarlo.Lattice — Type.

Abstract definition of a lattice. Necessary fields depend on Monte Carlo flavor. However, any concrete Lattice type should have at least the following fields:

- `sites`: number of lattice sites
- `neighs::Matrix{Int}`: neighbor matrix (row = neighbors, col = siteidx)

source

MonteCarlo.Model — Type.

Abstract definition of a model. A concrete model type must have two fields:

- `β::Float64`: temperature (depends on MC flavor if this will actually be used)
- `l::Lattice`: any [Lattice](@ref)

A concrete model must implement the following methods:

- `conftype(m::Model)`: type of a configuration
- `energy(m::Model, conf)`: energy of configuration
- `rand(m::Model)`: random configuration
- `propose_local(m::Model, i::Int, conf, E::Float64) -> ΔE, Δi`: propose local move
- `accept_local(m::Model, i::Int, conf, E::Float64)`: accept a local move

source

MonteCarlo.MonteCarloFlavor — Type.

Abstract definition of a Monte Carlo flavor.

A concrete monte carlo flavor must implement the following methods:

- `init!(mc)`: initialize the simulation without overriding parameters (will also automatically be available as `reset!`)
- `run!(mc)`: run the simulation

source