General

Methods: General

Below you find all general exports.

Index

MonteCarlo.observables

MonteCarlo.reset!

Documentation

MonteCarlo.observables — Method.

observables(mc::MonteCarloFlavor)

Get a list of all observables defined for a given Monte Carlo simulation.

Returns a Dict{String, String} where values are the observables names and keys are short versions of those names. The keys can be used to collect correponding observable objects from the Monte Carlo simulation, e.g. like mc.obs[key].

Note, there is no need to implement this function for a custom MonteCarloFlavor.

source

MonteCarlo.reset! — Method.

reset!(mc::MonteCarloFlavor)

Resets the Monte Carlo simulation mc. Previously set parameters will be retained.

source