General

Methods: General

Below you find all general exports.

Index

MonteCarlo.IsingModel

MonteCarlo.IsingModel

MonteCarlo.MC

MonteCarlo.MC

MonteCarlo.init!

MonteCarlo.observables

MonteCarlo.run!

Documentation

MonteCarlo.init! — Method.

init!(mc::MC[; seed::Real=-1])

Initialize the classical Monte Carlo simulation mc. If seed !=- 1 the random generator will be initialized with srand(seed).

source

MonteCarlo.run! — Method.

run!(mc::MC[; verbose::Bool=true, sweeps::Int, thermalization::Int])

Runs the given classical Monte Carlo simulation mc. Progress will be printed to STDOUT if verborse=true (default).

source

MonteCarlo.MC — Type.

Classical Monte Carlo simulation

source

MonteCarlo.MC — Method.

MC(m::M) where M<:Model

Create a classical Monte Carlo simulation for model m with default parameters.

source

MonteCarlo.IsingModel — Type.

Famous Ising model on a cubic lattice.

source

MonteCarlo.IsingModel — Method.

IsingModel(dims::Int, L::Int, β::Float64)
IsingModel(; dims::Int=2, L::Int=8, β::Float64=1.0)

Create Ising model on dims-dimensional cubic lattice with linear system size L and inverse temperature β.

source

MonteCarlo.observables — Method.

observables(m::Model)

Get all observables defined for a given model.

Returns a Dict{String, String} where values are the observables names and keys are short versions of those names. The keys can be used to collect correponding observable objects from a Monte Carlo simulation, e.g. like mc.obs[key].

source