Models

Methods: Models

Below you find all methods that any particular implementation of the abstract type MonteCarlo.Model should implement. See Custom models for more information.

Index

Base.Random.rand

MonteCarlo.accept_local!

MonteCarlo.conftype

MonteCarlo.energy

MonteCarlo.finish_observables!

MonteCarlo.global_move

MonteCarlo.measure_observables!

MonteCarlo.observables

MonteCarlo.prepare_observables

MonteCarlo.propose_local

Documentation

MonteCarlo.observables — Method.

observables(m::Model)

Get all observables defined for a given model.

Returns a Dict{String, String} where values are the observables names and keys are short versions of those names. The keys can be used to collect correponding observable objects from a Monte Carlo simulation, e.g. like mc.obs[key].

source

Base.Random.rand — Method.

rand(m::Model)

Draw random configuration.

source

MonteCarlo.accept_local! — Method.

accept_local(m::Model, i::Int, conf, E::Float64)

Accept a local move for site i of current configuration conf with energy E. Arguments Δi and ΔE correspond to output of propose_local() for that local move.

See also propose_local.

source

MonteCarlo.conftype — Method.

conftype(m::Model)

Returns the type of a configuration.

source

MonteCarlo.energy — Method.

energy(m::Model, conf)

Calculate energy of configuration conf for Model m.

source

MonteCarlo.finish_observables! — Method.

measure_observables!(m::Model, obs::Dict{String,Observable}, conf, E::Float64)

Measure observables and update corresponding MonteCarloObservable.Observable objects in obs.

See also prepare_observables and measure_observables!.

source

MonteCarlo.global_move — Method.

global_move(m::Model, conf, E::Float64) -> accepted::Bool

Propose a global move for configuration conf with energy E. Returns wether the global move has been accepted or not.

source

MonteCarlo.measure_observables! — Method.

measure_observables!(m::Model, obs::Dict{String,Observable}, conf, E::Float64)

Measures observables and updates corresponding MonteCarloObservable.Observable objects in obs.

See also prepare_observables and finish_observables!.

source

MonteCarlo.prepare_observables — Method.

prepare_observables(m::Model) -> Dict{String, Observable}

Initializes observables and returns a Dict{String, Observable}. In the latter, keys are abbreviations for the observables names and values are the observables themselves.

See also measure_observables! and finish_observables!.

source

MonteCarlo.propose_local — Method.

propose_local(m::Model, i::Int, conf, E::Float64) -> ΔE, Δi

Propose a local move for lattice site i of current configuration conf with energy E. Returns local move information Δi (e.g. new[i] - conf[i], will be forwarded to accept_local!) and energy difference ΔE = E_new - E_old.

See also accept_local!.

source